Fix conditional export
